+++ src/timeline/Timeline.js
@@ -17,6 +17,7 @@ function Timeline (container, items, options) {
         // zoomable: true, // TODO: option zoomable
         showMinorLabels: true,
         showMajorLabels: true,
+        showCurrentTime: false,
         autoResize: false
     }, options);
 
@@ -76,6 +77,14 @@ function Timeline (container, items, options) {
         start: now.clone().add('days', -3).valueOf(),
         end:   now.clone().add('days', 4).valueOf()
     });
+  /* TODO: fix range options
+    var rangeOptions = Object.create(this.options);
+    this.range = new Range(rangeOptions);
+    this.range.setRange(
+        now.clone().add('days', -3).valueOf(),
+        now.clone().add('days', 4).valueOf()
+    );
+    */
     // TODO: reckon with options moveable and zoomable
     this.range.subscribe(this.rootPanel, 'move', 'horizontal');
     this.range.subscribe(this.rootPanel, 'zoom', 'horizontal');
@@ -101,6 +110,10 @@ function Timeline (container, items, options) {
     this.timeaxis.setRange(this.range);
     this.controller.add(this.timeaxis);
 
+    // current time bar
+    this.currenttime = new CurrentTime(this.timeaxis, [], rootOptions);
+    this.controller.add(this.currenttime);
+
     // create itemset or groupset
     this.setGroups(null);
 
@@ -121,6 +134,8 @@ Timeline.prototype.setOptions = function (options) {
     if (options) {
         util.extend(this.options, options);
     }
+
+    // TODO: apply range min,max
 
     this.controller.reflow();
     this.controller.repaint();
@@ -164,6 +179,10 @@ Timeline.prototype.setItems = function(items) {
         var max = dataRange.max;
         if (min != null && max != null) {
             var interval = (max.valueOf() - min.valueOf());
+            if (interval <= 0) {
+                // prevent an empty interval
+                interval = 24 * 60 * 60 * 1000; // 1 day
+            }
             min = new Date(min.valueOf() - interval * 0.05);
             max = new Date(max.valueOf() + interval * 0.05);
         }