+++ HISTORY.md
@@ -1,6 +1,18 @@
 vis.js history
 http://visjs.org
 
+## <not yet released>, version 0.3.0
+
+- Implemented option `showCurrentTime`, displaying a red, vertical bar at
+  current time. Thanks fi0dor.
+
+
+## 2013-09-20, version 0.2.0
+
+- Implemented full touch support for Graph.
+- Fixed initial empty range in the Timeline in case of a single item.
+- Fixed field `className` not working for items.
+
 
 ## 2013-06-20, version 0.1.0